+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.3
 
 + Switched to `py-cord` framework, a fork of discord.py that does support subcommands.
-+ Added `/stats` command.
++ Added `/stats` and `/rich` commands.
++ Added multiple sources to `/lore` command.
 + `/handbook` available as a context-menu command on users.
 
 ## 0.2.1